restructure unittests to not use helper modules
Originally attempted to breakup the tests into individual modules; however, this resulted in module
not found errors during execution of the entire repo test suite. Combining all of the tests and
removing __init__ appears to resolve the issue.
