--- stix_shifter_modules/infoblox/test/stix_translation/__init__.py
@@ -1,1 +1,0 @@
-# -*- coding: utf-8 -*-
