--- stix_shifter_modules/infoblox/test/stix_transmission/__init__.py
@@ -1,1 +1,0 @@
-# -*- coding: utf-8 -*-
